requirements.txt
asgiref==3.7.2
Django==4.2.5
django-taggit==4.0.0
djangorestframework==3.14.0
importlib-metadata==6.8.0
Markdown==3.4.4
psycopg2==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
zipp==3.17.0
